requirements.txt
flask==3.1.0
python-dotenv==1.0.1
requests==2.32.3
python-docx==1.1.2
reportlab==4.2.5
moviepy==1.0.3
